rename seed_hex
